// src/types/components.ts
export const defaultFilters: FilterValues = {
  brand: "",
  category: [],
  pricemin: 0,
  pricemax: 50000,
  sort: "",
};
// ...
export interface FilterValues {
  brand: string;
  category: string[];
  pricemin: number;
  pricemax: number;
  sort: string;
}

export interface FilterBarProps {
  onFilterChange: (filters: FilterValues) => void;
  initialFilters?: FilterValues;
}